+++ CHANGELOG.md
@@ -12,7 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Predictions in the dashboard are now sorted by date
 - The top-n returned predictions are now based on top-n unique patients, not individual appointments
-- Added clinic name and clinic number to api
+- Added clinic name and clinic phone number to api
+- Dashboard now shows clinic name and the patient index
+
+### Fixed
+- Dashboard no longer crashes when trying to display non-existing sensitive patient info. It now shows a message, while still displaying the (anonymized) appointments.
 
 ## [1.1.2] - 2023-09-18